[wasm] Remove the old fuzzer corpus from gitignore
In a recent CL I moved the corpus of the wasm fuzzer and of the
wasm-asmjs fuzzer to a different directory
(wasm_corpus and wasm_asmjs_corpus) so that the corpus is not executed
on the try-bots. With this CL I remove the old corpus from the
.gitignore file. All *.wasm files can be deleted safely from the
directories test/fuzzer/wasm and test/fuzzer/wasm_asmjs.

R=bradnelson@chromium.org
CC=machenbach@chromium.org

Change-Id: I9fdf9d77b71b133f84f7e744763d65fdf127d624