Update patch set 3
Patch Set 3:

Yang: I'd like to get rid of the unnecessary check for an input being "observable" in the RegisterOptimizer. This means that the bytecode will use a "temporary" register for operations instead of the  "local" registers if it knows the two are equivalent. This impacts debugging of optimized code in that from TurboFan's point-of-view, the local registers are dead, and the values instead come from the temporary registers. This means that the debugger might not be able to see locals when debugging optimized code (although I think this could always a problem depending on what optimizations TF does, and this just makes it slightly worse).

I've updated the debug-evalutate tests to add a if/else branch after the call, which flushes the bytecode register optimizer, and keeps the tests passing. WDYT?

Patch-set: 3
Reviewer: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>