[compiler] Use smi zero for empty optimized code cell
WIP: x64-only

Instead of initialising the optimized code cell slot in the feedback
vector to an empty WeakCell, initialise it (and clear it) to a Smi zero
value, only creating the WeakCell when needed.

This is a potential precursor to storing other Smi-valued sentinels in
the optimized code sell slot.

Bug: v8:6246
Bug: v8:6360
Change-Id: I338ed42ad8d9e0c68a276e8e8056892aa482d203