Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Here's a merge of Dan's fix (rather than a revert). git cherry-pick was able to handle the rename OK. PTAL



For Chrome, just reverting is actually better (not just in branch but also in trunk) because the default locale is Chrome's UI locale and the ICU data bundled with Chrome supports all the UI locales although some of them are mistaken for not being supported due to the way getAvailabeLocale works. 

Sorry that I didn't make this point (more unambiguously). 

For v8-stand-alone, that's a different story because we don't have the above 'out-of-band' information. 

Is there a way to distinguish whether v8 is built as a part of Chrome or standalone? 

Even if there is, a branch would not be a good place to do that. So, let's go with this CL for 59 branch.  

In v8:6288, I'll think about what to do.

Patch-set: 1
Reviewer: Gerrit User 1151758 <1151758@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>