Update patch set 2
Patch Set 2:

ok, I guess this has some problems and I don't 100% know how to fix them atm.

It might be okay to leave the bug as it is since the process for getting rid of crankshaft seems to be speeding up, and I no longer believe this is reproducible with turbofan (as there is no situation where an eagerly parsed function will be re-parsed for optimization, if I'm not mistaken).

Patch-set: 2
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>