[wasm] Throttle the amount of unfinished work to avoid OOM
It is possible that the foreground task is unable to clear the
scheduled unfinished work, eventually leading to an OOM.

The solution here has a certain elasticity - we can actually have
almost 2 x max number of allowed unfinished work. This fits the goal
of throttling, and it keeps the change smaller. The "tighter control"
alternative would require a separate counter (nr of accepted work).

Bug:v8:6492
Change-Id: I1e4c0825351a42fa0b8369ccc41800ac3445563d