[wasm] Throttle the amount of unfinished work to avoid OOM
It is possible that the foreground task is unable to clear the
scheduled unfinished work, eventually leading to an OOM.

We use either code_range on 64 bit, or the capacity of the code space,
as a heuristic for how much memory to use for compilation.

Additionally, this change avoids keeping the main thread busy with 
anything else than code generation, since we want to clear that queue
as fast as possible, to avoid hitting the memory threshold.

Bug:v8:6492
Change-Id: I1e4c0825351a42fa0b8369ccc41800ac3445563d