Update patch set 18
Patch Set 18:

PTAL.

There appears to be a regression in compile time on x86 only. Since it's not happening on arm32, nor on any of the 64 bit systems, I'd prefer investigating separately and moving with this CL, to unblock back-merging it.

Patch-set: 18
Reviewer: Gerrit User 1125279 <1125279@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>