[TurboFan] Keep track of types in registers across suspend/resume.
Adds tracking of the nodes which are saved in the generator's register
file on suspend and adds these nodes to the corresponding register restores
in the resume. These nodes are then used to propegate typing information
across the suspend / resume, since the values don't change between
suspend and resume.

This is needed for StringConcat since that operator assumes all inputs are
types, and without tracking across suspend / resume then expressions like
a + yield + b will forget that the value for a is a string. It should
also enable generation of better code.

BUG=v8:6243

Change-Id: Iab18a08ff4c5df49b5a30e03f579dfe9f5ecb410