Update patch set 2
Patch Set 2:

FYI - The v8 roll when this landed was a 49kb binary size increase on Android, and the roll with this revert was a 49kb decrease.

If a reland is in the works, it would be great if you could also run it through //tools/binary_size/diagnose_bloat.py to see if there is any way to reduce its binary size impact.

Graph links:
https://chromeperf.appspot.com/report?sid=a097e74b1aa288511afb4cb616efe0f95ba4d347ad61d5e835072f23450938ba&rev=482007
https://chromeperf.appspot.com/report?sid=a097e74b1aa288511afb4cb616efe0f95ba4d347ad61d5e835072f23450938ba&rev=482605

Patch-set: 2
CC: Gerrit User 1126674 <1126674@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>