Update patch set 4
Patch Set 4: Code-Review+1

(6 comments)

Thanks for reworking this. I think it is an improvement over doing it in the ast-numbering phase, and since we are hoping to remove that phase entirely at some point this is a good step in that direction. LGTM if all comments can be addressed.

Patch-set: 4
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1