[Turbofan] Change the revisit order to breadth first instead of depth first.
When there are some changes in the graph, we mark the uses of the reduced
node for revisiting. These are pushed onto a stack and hence follow a
depth first traversal. This sometimes leads to propogating the changes
down the graph and again revisiting the entire graph because of the
nodes that were marked for revisit earlier.

Bug:chromium:725664
Change-Id: I036fd7096c422565a2851b109aba2b4c438403c0