[wasm] The module decoder should not own the WasmModule it generates
At the moment the module decoder owns the WasmModule while it does the
decoding, and gives ownership of the module to the caller after decoding
finished. In normal compilation, the ownership is passed to a Managed
object on the JS heap which is created during the initialization of the
function compilation. For streaming compilation, function compilation has
to be initialized before the decoding is actually finished. This means
that the module decoder cannot own the WasmModule until decoding is
finished, because ownership has to be transferred to the Managed object.

With this CL the module decoder operates on a raw pointer to a
WasmModule instead of a unique_ptr. Thereby we can take the ownership
of the WasmModule from the module decoder any time we want.

A side effect is that FinishDecoding cannot produce the actual decoding
result anymore, because for that FinishDecoding would require ownership
of the WasmModule. Instead it produces a dummy result, and the error
result from the dummy is used to create the actual result.

R=clemensh@chromium.org

Change-Id: Id0646c48a507874dc14afe1eeeed1691c7868bf9