[wasm] Limit the output length of user-provided names
We currently have a fixed limit of 256 characters for error messages
generated in the decoder. We want to keep a fixed limit, even if
user-provided names appear in the output. Hence, names longer than a
certain threshold have to be truncated.
This CL introduces macros to do this, and uses them for truncating the
name of exports in error messages. We might want to use it in more
locations in the future.

R=titzer@chromium.org

Bug: chromium:740023
Change-Id: I47d4125da943f50a01a5aa164f1f3c9e7a70a539