[Compiler] Remove HasAsmWasmData check from ShouldUseFullcodeGen.
We should never try to compile unoptimized code for a function with
AsmWasm data, since it either shouldn't have been created, or should be
cleared before we reach here, so remove the check.

Also change the check for optimizing using bytecode to be a check on
whether the function HasBytecode.

BUG=v8:5203, v8:6409

Change-Id: I472b123bb370f7929555e6162af02173834d01a1