Update patch set 2
Patch Set 2: Code-Review+1

Fine to land as is for now if we duplicate the benchmarks that use it and create extra config files that only have one line item (the process size).

This will create some extra load, dependent on how often we wanna use it. We can selectively choose it and e.g. not use it on Android.

Patch-set: 2
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1