Update patch set 2
Patch Set 2: Code-Review+1

Looks good to me if Jaro and Benedikt are happy. Please followup with a CL to rename it and move it to runtime-test.cc to make it clear it's just for tests if the CL sticks.

Patch-set: 2
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1