Update patch set 1
Patch Set 1:

(1 comment)

I had a look at this with Jaro, and it seems like the problem is that the IndexOf/Includes descriptors specify IntPtr types for length/fromIndex arguments. The error we're running into here occurs because the compiler ends up reusing the length node for both the call to IndexOf (IntPtr type) and a StoreField (expecting a signed tagged value).

The compiler doesn't know how to handle this combination of types, therefore the error.

At least that's my understanding of the issue; Jaro will know more.

The short-term fix is to change the stubs to take an Int32 or Tagged types. Longer-term, we could teach the compiler to handle this case.

Personally, I'm not sure we gain that much by passing raw ints. Maybe we can get away with passing around Smis and untag when needed (then we could also remove the custom interface descriptors and use normal TFS stubs).

And finally, how to debug this:

--trace-turbo-graph will print the generated graph,
--trace-representation representation selection.

We're failing here: 

   queue #47?: truncate-to-word64

and searching for #47 gives us the use sites

#48:StoreField[tagged base, 24, Range(0, 134217726), kRepTaggedSigned|kTypeInt32, NoWriteBarrier](#42:Allocate, #47:NumberConstant, #46:StoreField, #12:JSStackCheck)

#53:Call[Code:ArrayIndexOf Descriptor:r1s0i6f0t0](#52:HeapConstant, #49:FinishRegion, #39:HeapConstant, #26:NumberAdd, #47:NumberConstant,
#25:NumberConstant, #24:SpeculativeNumberSubtract, #12:JSStackCheck)  [Type: Any]

Patch-set: 1