Update patch set 5
Patch Set 5:

(5 comments)

Thanks for the look. I agree it's hard to follow everything that goes on in generators (especially async generators).

The flow is something like:

A) when generator .next()|.return()|.throw() methods are called, append a new AsyncGeneratorRequest to the generator's queue.
B) if the generator is not currently executing, perform AsyncGeneratorResumeNext.
C) AsyncGeneratorResumeNext: Repeat while there are requests in the queue, and the generator is not suspended for an Await:
    i. Let request be the first request in the queue
    ii. If the request has a resume mode that is "return" or "throw",
        ii.1: close the generator if it has not yet been started
        ii.2: if resume mode is "return",
             ii.2.a) Suspend the generator, recording that it is suspended
                     for Await (such that AsyncGeneratorResumeNext does not
                     process following items in the queue)
             ii.2.b) Let result be Await(request.value)
             ii.2.c) Record that the generator is no longer suspended for Await.
             ii.2.d) If the generator is closed,
                 ii.2.d.1) If result is an abrupt completion, perform
                           AsyncGeneratorReject(generator, result.value)
                 ii.2.d.2) Else perform AsyncGeneratorResolve(generator,
                               result.value, true)
             ii.2.e) Else if result is an abrupt completion, resume the
                     generator with resume mode "throw"
             ii.2.e) Else resume the generator with resume mode "return"
         ii.3: Else, assert resume mode is "throw"
             ii.3.a) If the generator is closed, perform
                     AsyncGeneratorReject(generator, request.value)
             ii.3.b) Else resume the generator with resume mode "throw"
    iii. Else, assert resume mode is "next".
        iii.a) Resume the generator with resume mode "next"

D) AsyncGeneratorResolve and AsyncGeneratorReject both remove the first rqeuest from the generator's queue of requests, and resolve or reject that request's promise with a particular value. In the case of AsyncGeneratorResolve, that value (and a boolean "done") are converted into an Iterator Result object.

E) AsyncGeneratorYield:
    i. Suspend the generator, recording that it is suspended for Await (such
       that AsyncGeneratorResumeNext does not process following items in the
       queue)
    ii. Let result be Await(value)
    iii. Record that the generator is no longer suspended for an Await.
    iv. If result is an abrupt completion, resume the generator with resume 
         mode "throw"
    v. Else, perform AsyncGeneratorResolve(result.value, false), and process
         remaining requests in the generator's queue (tail-call to
         AsyncGeneratorResumeNext)

F) AsyncGeneratorAwait:
    i. Suspend the generator, recording that it is suspended for Await (such
       that AsyncGeneratorResumeNext does not process following items in the
       queue)
    ii. Let result be Await (value).
    iii. Record that the generator is no longer suspended for an Await.
    iv. If result is an abrupt completion, resume the generator with resume 
         mode "throw"
    v. Else resume the generator with resume mode "next"


---

When the generator is resumed with mode "throw", it will immediately throw an exception after jumping to the right PC (such that the correct handler table
is used to handle the exception, if any.)

When the generator is resumed with mode "return", it will eventually try to
return, evaluating any finally blocks along the way.

Otherwise, the generator continues normally, immediately following the last suspend.

----

So, descriptions like this are still a bit hard to follow. Should I include something like this as a comment somewhere in async-generator-builtins-gen.cc? I don't know that it really helps with anyone's mental model of how it works :(

Patch-set: 5
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0