Update patch set 2
Patch Set 2:

(1 comment)

Thanks, Jakob (and Jaro)!

The --trace* options are very interesting, although still a bit challenging for me to read. :)

If I went with the custom descriptor and using Int32, do I need to modify builtins-array-gen.cc as well? The code currently uses IntPtr, which I understand can be 32 or 64 bits. I actually tried to rewrite it to use Int32, but then I triggered a DCHECK in CodeStubAssembler::Increment and realised that the input to Increment (called on the |index_var|) needs to be either IntPtr or SMI.

Which brings me to your suggestion of switching to TFS and passing SMIs around. I think I know how to implement that (delete the custom descriptor, redeclare the builtins as TFS, use the parameters from the Descriptor:: class/namespace). Given that it simplifies the code, I'm happy to go for it first and deal with potential speed regressions if they appear.

The current patchset is still not done. The optimised version of indexOf for some reason reports "not found" (-1) on [0].indexOf(0), and I will continue investigating it.

I will keep this CL updated. Thanks for the help so far.

Cheers,
Vaclav

Patch-set: 2
Reviewer: Gerrit User 1002541 <1002541@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>