[modules] Fix mapping of internal status to external status.
Internally, the module status type has one additional value, namely
PreInstantiating. I previously mapped this to Instantiating when
crossing the API boundary but it really should be mapped to
Uninstantiated. That's because when instantiation fails, typically all
modules not yet visited will remain in the PreInstantiating state, yet
they must appear Uninstantiated to the outside.

Bug: chromium:748544
Change-Id: Icb33c7f90db5e62375b6c09d14e3d2d5342b0879