Update patch set 8
Patch Set 8: Commit-Queue+1

(10 comments)

Thanks for the reviews. Comments addressed.

I've started writing a design doc, but it's not finished yet. The basic idea is simple: instead of storing {a WeakCell pointing to a target map} directly, we store {a handler that among other things contains {a WeakCell pointing to a target map}}, so that we can look up the cached handler when we need it again instead of having to recompute it. Callers requiring the target map simply "reach through" the handler to retrieve it.

Patch-set: 8
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1