[wasm] Introduce the Wasm Context struct as global External Reference
The WasmContext struct is used to store the mem_size and mem_start address of
the wasm instance linear memory. The code then uses these values to generate
ExternalReference TF nodes when these values are needed in the generated code
(i.e., access boundary checks).

When the memory changes (instance initialization, reset, grow_memory), then the
context variable are changed, so that the generated code will load the correct
values.  is declared in src/assembler.h as extern and defined in
src/compiler/wasm-compiler.cc.

In this prototype, the wasm_context variable is defined globally (in future
prototypes, the context needs to be associated to the instance, or even better
to the memory object); so the tests that leverage multiple instances are
failing.


Details for failing tests:

 * mjsunit/wasm/instantiate-module-basic

     Reason: multiple instances.

 * mjsunit/wasm/grow-memory

     Reason: the previous code relied on integer loop around for the boundary
     checks since the compuation was made directly at C level. Now that the
     compuation is made at TF level, this needs some reworking.

 * mjsunit/wasm/compiled-module-serialization

     Reason: multiple instances.

 * mjsunit/wasm/instance-gc:

     Reason: multiple instances.
Change-Id: I0b953a53cbfdd9c49cbb4b0f1fc57ddc3726e787